[CLOUD-2408] Added a routes.go file which contains type of routes values.yaml
